reports_service/requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
Flask-JWT-Extended
Werkzeug
python-dotenv